% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancedKmeans.R
\name{balanced_kmeans}
\alias{balanced_kmeans}
\title{The balanced k-means algorithm}
\usage{
balanced_kmeans(x, k = 2, iter.max = 10)
}
\arguments{
\item{x}{a matrix or data frame to be clustered, each row represents an item}

\item{k}{number of clusters}

\item{iter.max}{maximum number of iteration}
}
\value{
A list of following items:
cluster: cluster numbers of the each item
centers: the centroids
size: size of the clusters
iter: final number of iteration
}
\description{
The balanced k-means algorithm
}
